n=int(input()) print("%.2f"%(3.14*n*n))

+++ Area_of_a_circle.py
@@ -1,2 +1,2 @@
-n=float(input())
+n=int(input())
 print("%.2f"%(3.14*n*n))
